melhorando delete

+++ src/main/java/com/javanfood/javanfood/domain/service/CadastroCozinhaService.java
@@ -1,5 +1,6 @@
 package com.javanfood.javanfood.domain.service;
 
+import com.javanfood.javanfood.domain.repository.CidadeRepository;
 import com.javanfood.javanfood.domain.repository.CozinhaRepository;
 import com.javanfood.javanfood.domain.exeption.EntidadeEmUsoExeption;
 import com.javanfood.javanfood.domain.exeption.EntidadeNaoEncontradaExeption;
@@ -12,22 +13,31 @@ import org.springframework.stereotype.Service;
 @Service
 public class CadastroCozinhaService {
 
-    @Autowired
-    private CozinhaRepository cozinhaRepository;
+	private final CidadeRepository cidadeRepository;
 
-    public Cozinha salvar(Cozinha cozinha) {
-        return cozinhaRepository.save(cozinha);
-    }
+	@Autowired
+	private CozinhaRepository cozinhaRepository;
 
-    public void excluir(Long cozinhaId) {
-        try {
-            cozinhaRepository.deleteById(cozinhaId);
-        } catch (EmptyResultDataAccessException e) {
-            throw new EntidadeNaoEncontradaExeption("Não existe cadastro de cozinha com código: " + cozinhaId);
-        } catch (DataIntegrityViolationException e) {
-            throw new EntidadeEmUsoExeption
-                    ("Cozinha de código " + cozinhaId + " não pode ser removida. Está em uso");
-        }
+	CadastroCozinhaService(CidadeRepository cidadeRepository) {
+		this.cidadeRepository = cidadeRepository;
+	}
 
-    }
+	public Cozinha salvar(Cozinha cozinha) {
+		return cozinhaRepository.save(cozinha);
+	}
+
+	public void excluir(Long cozinhaId) {
+
+		if (!cozinhaRepository.existsById(cozinhaId)) {
+			throw new EntidadeNaoEncontradaExeption("Não existe cadastro de cozinha com código: " + cozinhaId);
+		}
+
+		try {
+			cozinhaRepository.deleteById(cozinhaId);
+			cozinhaRepository.flush();
+		} catch (DataIntegrityViolationException e) {
+			throw new EntidadeEmUsoExeption("Cozinha de código " + cozinhaId + " não pode ser removida. Está em uso");
+		}
+
+	}
 }
